% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdMold.R
\name{mdMoldViewServer}
\alias{mdMoldViewServer}
\title{处理逻辑}
\usage{
mdMoldViewServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
mdMoldViewServer()
}
